requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
joblib==1.3.2
kaleido==0.2.1
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mysqlclient==2.2.4
namex==0.0.7
numpy==1.26.3
opt-einsum==3.3.0
optree==0.10.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.20.0
protobuf==4.25.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
wheel==0.41.2
whitenoise==6.6.0
wrapt==1.16.0
